[+] Switched over to using the php5 version of ADOdb. (http://internap.dl.sourceforge.net/sourceforge/adodb/adodb500beta.tgz)
The one caveat is that adodb-session2.php was not updated in their beta version.
I did the quick fix version (rather than rewriting their entire class properly for php5)
and left the original version, a unified diff, and the patched version inside the directory.

If it is found that something is broke with this version (seems to work for me), then to change back, all you need to do is edit

devblocks/api/Engine.php

and change the includes from libs/adodb5/.... to be libs/adodb/ (it occurs twice in that file)



git-svn-id: http://svn.webgroupmedia.com/devblocks/devblocks@72 747e5740-6425-0410-825b-f4fe3d30fc61